requirements.txt
Django>=4.2.0
openpyxl>=3.1.0
Pillow>=10.0.0
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.0.0
reportlab>=4.0.0
xhtml2pdf>=0.2.11
celery>=5.3.0
redis>=5.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8
django-allauth>=0.57.0
channels>=4.0.0
daphne>=4.0.0
